refactor: src db
存在严重问题，重新设计架构和代码细节
